requirements.txt
# Core data & ML libs
numpy>=1.23.0,<1.27.0
pandas>=1.4.0,<2.0.0
scikit-learn>=1.0.0,<1.3.0
torch
torchvision
math

# Visualization & progress
tqdm>=4.62.3
matplotlib>=3.5.1
seaborn>=0.11.2
Pillow>=9.0.1
imageio>=2.14.1

# DICOM support
pydicom>=2.0.0

# Model / infra
transformers>=4.12.3

# API & UI
streamlit>=1.24.0
fastapi>=0.95.0
uvicorn[standard]>=0.22.0

# Utilities & testing
kaggle>=1.5.12
pytest>=6.2.5
pytest-cov>=3.0.0
openpyxl>=3.1.5
python-multipart>=0.0.5
requests>=2.28.0